Add generate to import .scss-lint.yml easily
While it has some quirks, which mainly have to be fixed on roger side of
things, this gives an nice idea of how easy (read: no excuse) it is to add a
config. Updated the documents to reflect usages and made adding a project
specific .scss-lint.yml file a recommendation.